+++ config/projects/gitlab.rb
@@ -15,7 +15,7 @@
 ## limitations under the License.
 ##
 #
-ee = system("#{Omnibus::Config.project_root}/support/is_gitlab_ee.sh") || system("#{Omnibus::Config.project_root}/support/is_gitlab_com.sh")
+ee = system("#{Omnibus::Config.project_root}/support/is_gitlab_ee.sh")
 
 if ee
   name "gitlab-ee"
@@ -44,20 +44,23 @@ build_iteration 1
 
 override :ruby, version: '2.1.6',  source: { md5: "6e5564364be085c45576787b48eeb75f" }
 override :rubygems, version: '2.2.1'
-override :'chef-gem', version: '11.18.0'
-override :'omnibus-ctl', version: '0.3.3'
+override :chef, version: '12.4.0.rc.0'
+override :'omnibus-ctl', version: '0.3.4'
 override :zlib, version: '1.2.8'
+override :cacerts, version: '2015.04.22', source: { md5: '380df856e8f789c1af97d0da9a243769' }
+override :redis, version: '2.8.20', source: { md5: 'a2588909eb497719bbbf664e6364962a' }
 
 # Openssh needs to be installed
 runtime_dependency "openssh-server"
 
 # creates required build directories
 dependency "preparation"
+dependency "package-scripts"
 
 dependency "git"
 dependency "redis"
 dependency "nginx"
-dependency "chef-gem"
+dependency "chef"
 dependency "remote-syslog" if ee
 dependency "logrotate"
 dependency "runit"
@@ -76,10 +79,10 @@ dependency "version-manifest"
 exclude "\.git*"
 exclude "bundler\/git"
 
-# Because we have a dynamic 'name' (gitlab-ce or gitlab-ee), omnibus-ruby would
-# look in either package-scripts/gitlab-ce or package-scripts/gitlab-ee. We
-# don't want that so let's hard-code the path.
-package_scripts_path "#{Omnibus::Config.project_root}/package-scripts/gitlab"
+# Our package scripts are generated from .erb files,
+# so we will grab them from an excluded folder
+package_scripts_path "#{install_dir}/.package_util/package-scripts"
+exclude '.package_util'
 
 package_user 'root'
 package_group 'root'