+++ files/gitlab-cookbooks/gitlab/libraries/gitlab.rb
@@ -109,7 +109,7 @@ module Gitlab
       unless uri.host
         raise "GitLab external URL must include a schema and FQDN, e.g. http://gitlab.example.com/"
       end
-      Gitlab['user']['git_user_email'] ||= "gitlab@#{uri.host}"
+
       Gitlab['gitlab_rails']['gitlab_host'] = uri.host
       Gitlab['gitlab_rails']['gitlab_email_from'] ||= "gitlab@#{uri.host}"
 
@@ -174,6 +174,33 @@ module Gitlab
         gitlab_rails.values_at('redis_host', 'redis_port') == gitlab_ci.values_at('redis_host', 'redis_port')
         Chef::Log.warn "gitlab-rails and gitlab-ci are configured to connect to "\
                        "the same Redis instance. This is not recommended."
+      end
+    end
+
+    def parse_postgresql_settings
+      # If the user wants to run the internal Postgres service using an alternative
+      # DB username, host or port, then those settings should also be applied to
+      # gitlab-rails and gitlab-ci.
+      [
+        # %w{gitlab_rails db_username} corresponds to
+        # Gitlab['gitlab_rails']['db_username'], etc.
+        [%w{gitlab_rails db_username}, %w{postgresql sql_user}],
+        [%w{gitlab_rails db_host}, %w{postgresql listen_address}],
+        [%w{gitlab_rails db_port}, %w{postgresql port}],
+        [%w{gitlab_ci db_username}, %w{postgresql sql_ci_user}],
+        [%w{gitlab_ci db_host}, %w{postgresql listen_address}],
+        [%w{gitlab_ci db_port}, %w{postgresql port}],
+      ].each do |left, right|
+        if ! Gitlab[left.first][left.last].nil?
+          # If the user explicitly sets a value for e.g.
+          # gitlab_rails['db_port'] in gitlab.rb then we should never override
+          # that.
+          next
+        end
+
+        better_value_from_gitlab_rb = Gitlab[right.first][right.last]
+        default_from_attributes = node['gitlab'][right.first.gsub('_', '-')][right.last]
+        Gitlab[left.first][left.last] = better_value_from_gitlab_rb || default_from_attributes
       end
     end
 
@@ -263,6 +290,7 @@ module Gitlab
       parse_git_data_dir
       parse_udp_log_shipping
       parse_redis_settings
+      parse_postgresql_settings
       parse_nginx_listen_address
       # Parse ci_external_url _before_ gitlab_ci settings so that the user
       # can turn on gitlab_ci by only specifying ci_external_url